fine controllers

+++ src/main/java/amintabite/U5_W3_D5/Services/EventoService.java
@@ -3,10 +3,15 @@ package amintabite.U5_W3_D5.Services;
 import amintabite.U5_W3_D5.Entities.Evento;
 import amintabite.U5_W3_D5.Entities.Utente;
 import amintabite.U5_W3_D5.Exceptions.NotFoundException;
+import amintabite.U5_W3_D5.Exceptions.UnauthorizedException;
 import amintabite.U5_W3_D5.Payloads.EventoPayload;
 import amintabite.U5_W3_D5.Repositories.EventoRepository;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -17,6 +22,15 @@ import java.util.UUID;
 public class EventoService {
     @Autowired
     private EventoRepository eventoRepository;
+
+
+    public Page<Evento> findAll(int page, int size, String sortBy) {
+        if (size > 20) size = 20; // limite max
+        Pageable pageable = PageRequest.of(page, size, Sort.by(sortBy));
+        return eventoRepository.findAll(pageable);
+    }
+
+
 
     public Evento saveEvento(EventoPayload payload, Utente organizzatore) {
         Evento evento = new Evento(
@@ -39,8 +53,17 @@ public class EventoService {
                 .orElseThrow(() -> new NotFoundException("Evento non trovato"));
     }
 
-    public Evento updateEvento(UUID eventoId, EventoPayload payload) {
-        Evento found = findById(eventoId);
+    public Evento updateEvento(UUID eventoId, EventoPayload payload ,Utente currentUser) {
+
+        Evento found = eventoRepository.findById(eventoId)
+                .orElseThrow(() -> new NotFoundException("Evento non trovato"));
+
+        if (!found.getOrganizzatore().getUtenteId().equals(currentUser.getUtenteId())) {
+            throw new UnauthorizedException("Non puoi modificare questo evento");
+        }
+
+
+
         found.setNPosti(payload.nPosti());
         found.setTitolo(payload.titolo());
         found.setDataEvento(payload.dataEvento());
@@ -49,9 +72,15 @@ public class EventoService {
         return eventoRepository.save(found);
     }
 
-    public void deleteEvento(UUID eventoId) {
+    public void deleteEvento(UUID eventoId, Utente currentUser) {
+
         Evento found = findById(eventoId);
+
+        if (!found.getOrganizzatore().getUtenteId().equals(currentUser.getUtenteId())) {
+            throw new UnauthorizedException("Non puoi eliminare questo evento");
+        }
+
         eventoRepository.delete(found);
-        log.info("Evento  eliminato correttamente", eventoId);
+        log.info("Evento  eliminato correttamente", found.getTitolo());
     }
 }
